requirements.txt
Django==2.2.10
django-filter==2.2.0
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
